dev/navigation: move package location
