+++ src/main/java/org/somerville/swag/display/Basket.java
@@ -1,4 +1,4 @@
-package org.somerville.swag.data.display;
+package org.somerville.swag.display;
 
 import org.somerville.swag.data.entities.Customer;
 import org.somerville.swag.data.entities.Order;
@@ -6,21 +6,14 @@ import org.somerville.swag.data.entities.Order;
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.util.Currency;
 
-public class LogIn {
+public class Basket {
+    private JTable tblBasket;
     public JPanel root;
-    private JTextField textField2;
-    private JTextField textField1;
-    private JButton confirmButton;
+    private JButton buyNowButton;
     private JButton backButton;
 
-    public LogIn(JFrame oldframe, Customer customer, Order order) {
-
-        if(customer != null){
-            confirmButton.setText("Sign Out");
-        }
-
+    public Basket(JFrame oldframe, Customer customer, Order order) {
         backButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
@@ -32,19 +25,18 @@ public class LogIn {
                 SwingUtilities.getWindowAncestor(root).dispose();
             }
         });
-        confirmButton.addActionListener(new ActionListener() {
+
+        buyNowButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                /**
-                 * needs to call create customer logic
-                 */
-                if(!customer.equals(null)){
-                    customer.equals(null);
-                }
-                else {
-                    //customer = *todologic*
-                }
+                JFrame frame = new JFrame("Somerville Swag");
+                frame.setContentPane(new Purchase(oldframe, customer, order).root);
+                frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+                frame.pack();
+                frame.setVisible(true);
+                SwingUtilities.getWindowAncestor(root).dispose();
             }
         });
     }
+
 }
